+++ index.html
@@ -231,6 +231,5 @@
     </div>
   </template>
   <script src="./js/main.js" type="module"></script>
-  <script src="./js/template.js" type="module"></script>
 </body>
 </html>
