+++ js/template.js
@@ -1,5 +1,6 @@
-import {createAdvertisements} from './main.js';
+import {createAdvertisements} from './advertisement.js';
 import {fillNodeTextContentOrHide} from './util.js';
+import {PROPERTY_NAMES, FEATURES_ICONS} from './data.js';
 
 const mapElement = document.querySelector('.map__canvas');
 const cardTemplate = document.querySelector('#card').content.querySelector('.popup');
@@ -7,23 +8,6 @@ const cardTemplate = document.querySelector('#card').content.querySelector('.pop
 const cards = createAdvertisements();
 
 const mapFragment = document.createDocumentFragment();
-
-const PROPERTY_NAMES = {
-  'flat' : 'Квартира',
-  'bungalow' : 'Бунгало',
-  'house' : 'Дом',
-  'palace' : 'Дворец',
-  'hotel' : 'Отель',
-};
-
-const FEATURES_ICONS = {
-  'wifi' : 'popup__feature--wifi',
-  'dishwasher' : 'popup__feature--dishwasher',
-  'parking' : 'popup__feature--parking',
-  'washer' : 'popup__feature--washer',
-  'elevator' : 'popup__feature--elevator',
-  'conditioner' : 'popup__feature--conditioner',
-};
 
 cards.forEach(({offer, author}) => {
   const cardElement = cardTemplate.cloneNode(true);
@@ -36,13 +20,6 @@ cards.forEach(({offer, author}) => {
   fillNodeTextContentOrHide(cardElement, '.popup__text--capacity', [offer.rooms, offer.guests], `${offer.rooms} комнаты для ${offer.guests} гостей`);
   fillNodeTextContentOrHide(cardElement, '.popup__text--time', [offer.checkin, offer.checkout],`Заезд после ${offer.checkin}, выезд до ${offer.checkout}`);
   fillNodeTextContentOrHide(cardElement, '.popup__description', offer.description);
-  // cardElement.querySelector('.popup__text--address').textContent = offer.address;
-  // cardElement.querySelector('.popup__text--price').textContent = `${offer.price} ₽/ночь`;
-  // cardElement.querySelector('.popup__type').textContent = PROPERTY_NAMES[offer.type];
-  // cardElement.querySelector('.popup__text--capacity').textContent = `${offer.rooms} комнаты для ${offer.guests} гостей`;
-  // cardElement.querySelector('.popup__text--time').textContent = `Заезд после ${offer.checkin}, выезд до ${offer.checkout}`;
-  // cardElement.querySelector('.popup__description').textContent = offer.description;
-
 
   const featureElements = cardElement.querySelectorAll('.popup__feature');
   for(let i = 0; i < featureElements.length; i++) {
